fix: refactor analytics.rs
