+++ sentry/src/db.rs
@@ -13,6 +13,7 @@ mod validator_message;
 
 pub mod channel;
 // mod channel;
+pub mod analytics;
 
 pub use self::channel::*;
 pub use self::validator_message::*;
